Prep for deb release as 2.70
This is a general tidyup to detach this fork better from Gordon, who was (and perhaps still is) completely inundated with WiringPi-related support questions.

This changeset includes a number of tweaks which should reduce support burden on Gordon. I believe they are essential if we plan to release a new debian package.

They should - hopefully - not reduce or make ambiguous the origins of this code while achieving the above.

* Remove Gordon's email address from all source files to avoid support queries finding their way to him
* Update wiringPi URLs to the GitHub repository to try and keep bugs/issues within the fork
* Bump version to 2.70 ready for a debian package release
* Update debian package maintainer info, add bugs & homepage URLs to this repository